Fix: LOSTVectorGenerationAlgorithm (#39)
# Description
The `LOSTVectorGenerationAlgorithm` is fixed, and successfully closes
#38. What we did:
- Identified the root cause (inconsistent conjugation of `orientation`
before rotation)
- Updated all test cases to reflect this
- ~Updated the Quaternion math to use forward quaternions instead~

Note: The diagnosis in #38 was backwards

This bug was a result of my negligence and misunderstanding of how the
native `Quaternion` interacts with `SphericalToQuaternion`, which
produces backwards rotations instead of forward rotations.

To mitigate this issue in the future, documentation has been added to
clearly state how that functionality works, with many redundancies.

# Artifacts for PR #39 (DO NOT CHANGE)
- [Coverage
Artifact](https://uwcubesat.github.io/found/39/merge/coverage)
- [Doxygen Artifact](https://uwcubesat.github.io/found/39/merge/doxygen)